refaktoring consumer
