+++ src/test/kotlin/no/nav/familie/ba/mottak/hendelser/JournalføringHendelseConsumerTest.kt
@@ -3,13 +3,9 @@ package no.nav.familie.ba.mottak.hendelser
 import io.mockk.*
 import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
-import no.nav.familie.ba.mottak.config.FeatureToggleService
+import no.nav.familie.ba.mottak.domene.HendelseConsumer
+import no.nav.familie.ba.mottak.domene.Hendelseslogg
 import no.nav.familie.ba.mottak.domene.HendelsesloggRepository
-import no.nav.familie.ba.mottak.integrasjoner.*
-import no.nav.familie.ba.mottak.task.OppdaterOgFerdigstillJournalpostTask
-import no.nav.familie.ba.mottak.task.OpprettOppgaveForJournalføringTask
-import no.nav.familie.prosessering.domene.Task
-import no.nav.familie.prosessering.domene.TaskRepository
 import no.nav.joarkjournalfoeringhendelser.JournalfoeringHendelseRecord
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.assertj.core.api.Assertions.assertThat
@@ -24,17 +20,13 @@ class JournalføringHendelseConsumerTest {
     @MockK(relaxed = true)
     lateinit var mockHendelsesloggRepository: HendelsesloggRepository
 
-    @MockK
-    lateinit var mockJournalpostClient: JournalpostClient
 
     @MockK(relaxed = true)
-    lateinit var mockTaskRepository: TaskRepository
+    lateinit var mockJournalhendelseService: JournalhendelseService
 
     @MockK(relaxed = true)
     lateinit var ack: Acknowledgment
 
-    @MockK(relaxed = true)
-    lateinit var mockFeatureToggleService: FeatureToggleService
 
     @InjectMockKs
     lateinit var consumer: JournalføringHendelseConsumer
@@ -43,214 +35,122 @@ class JournalføringHendelseConsumerTest {
     internal fun setUp() {
         MockKAnnotations.init(this)
         clearAllMocks()
-
-        //Inngående papirsøknad, Mottatt
-        every {
-            mockJournalpostClient.hentJournalpost(JOURNALPOST_PAPIRSØKNAD)
-        } returns Journalpost(journalpostId = JOURNALPOST_PAPIRSØKNAD,
-                              journalposttype = Journalposttype.I,
-                              journalstatus = Journalstatus.MOTTATT,
-                              bruker = Bruker("123456789012", BrukerIdType.AKTOERID),
-                              tema = "BAR",
-                              kanal = "SKAN_NETS",
-                              behandlingstema = null,
-                              dokumenter = null,
-                              journalforendeEnhet = null,
-                              sak = null)
-
-        //Inngående digital, Mottatt
-        every {
-            mockJournalpostClient.hentJournalpost(JOURNALPOST_DIGITALSØKNAD)
-        } returns Journalpost(journalpostId = JOURNALPOST_DIGITALSØKNAD,
-                              journalposttype = Journalposttype.I,
-                              journalstatus = Journalstatus.MOTTATT,
-                              bruker = Bruker("123456789012", BrukerIdType.AKTOERID),
-                              tema = "BAR",
-                              kanal = "NAV_NO",
-                              behandlingstema = null,
-                              dokumenter = null,
-                              journalforendeEnhet = null,
-                              sak = null)
-
-        //Utgående digital, Mottatt
-        every {
-            mockJournalpostClient.hentJournalpost(JOURNALPOST_UTGÅENDE_DOKUMENT)
-        } returns Journalpost(journalpostId = JOURNALPOST_UTGÅENDE_DOKUMENT,
-                              journalposttype = Journalposttype.U,
-                              journalstatus = Journalstatus.MOTTATT,
-                              bruker = Bruker("123456789012", BrukerIdType.AKTOERID),
-                              tema = "BAR",
-                              kanal = "SKAN_NETS",
-                              behandlingstema = null,
-                              dokumenter = null,
-                              journalforendeEnhet = null,
-                              sak = null)
-
-        //Ikke barnetrygd
-        every {
-            mockJournalpostClient.hentJournalpost(JOURNALPOST_IKKE_BARNETRYGD)
-        } returns Journalpost(journalpostId = JOURNALPOST_IKKE_BARNETRYGD,
-                              journalposttype = Journalposttype.U,
-                              journalstatus = Journalstatus.MOTTATT,
-                              bruker = Bruker("123456789012", BrukerIdType.AKTOERID),
-                              tema = "FOR",
-                              kanal = "NAV_NO",
-                              behandlingstema = null,
-                              dokumenter = null,
-                              journalforendeEnhet = null,
-                              sak = null)
-
-        //ferdigstilt journalpost
-        every {
-            mockJournalpostClient.hentJournalpost(JOURNALPOST_FERDIGSTILT)
-        } returns Journalpost(journalpostId = JOURNALPOST_FERDIGSTILT,
-                              journalposttype = Journalposttype.U,
-                              journalstatus = Journalstatus.FERDIGSTILT,
-                              bruker = Bruker("123456789012", BrukerIdType.AKTOERID),
-                              tema = "FOR",
-                              kanal = "NAV_NO",
-                              behandlingstema = null,
-                              dokumenter = null,
-                              journalforendeEnhet = null,
-                              sak = null)
-
-        every { mockFeatureToggleService.isEnabled(any()) } returns true
     }
 
     @Test
-    fun `Mottak av papirsøknader skal opprette OpprettOppgaveForJournalføringTask`() {
-        val record = opprettRecord(JOURNALPOST_PAPIRSØKNAD)
-
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
+    fun `Skal ignorere hendelse fordi den eksisterer i hendelseslogg`() {
+        val consumerRecord = ConsumerRecord("topic", 1, OFFSET, HENDELSE_ID, opprettRecord())
+        every {
+            mockHendelsesloggRepository.existsByHendelseIdAndConsumer(HENDELSE_ID.toString(), HendelseConsumer.JOURNAL)
+        } returns true
 
         consumer.listen(consumerRecord, ack)
 
-
-        val taskSlot = slot<Task>()
-        verify {
-            mockTaskRepository.save(capture(taskSlot))
+        verify(exactly = 0) {
+            mockJournalhendelseService.behandleJournalhendelse(any())
         }
-
-        assertThat(taskSlot.captured).isNotNull
-        assertThat(taskSlot.captured.payload).isEqualTo(JOURNALPOST_PAPIRSØKNAD)
-        assertThat(taskSlot.captured.metadata.getProperty("callId")).isEqualTo("kanalReferanseId")
-        assertThat(taskSlot.captured.taskStepType).isEqualTo(OpprettOppgaveForJournalføringTask.TASK_STEP_TYPE)
         verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
-    }
 
+        verify(exactly = 0) {
+            mockHendelsesloggRepository.save(any())
+        }
+    }
 
     @Test
-    fun `Mottak av digital søknader skal opprette task`() {
-        val record = opprettRecord(JOURNALPOST_DIGITALSØKNAD)
-
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
+    fun `Mottak av gyldig hendelse skal delegeres til service`() {
+        val consumerRecord = ConsumerRecord("topic", 1, OFFSET, HENDELSE_ID, opprettRecord())
 
         consumer.listen(consumerRecord, ack)
 
-        val taskSlot = slot<Task>()
         verify {
-            mockTaskRepository.save(capture(taskSlot))
+            mockJournalhendelseService.behandleJournalhendelse(any())
+        }
+        verify { ack.acknowledge() }
+
+        val slot = slot<Hendelseslogg>()
+        verify(exactly = 1) {
+            mockHendelsesloggRepository.save(capture(slot))
         }
 
-        assertThat(taskSlot.captured).isNotNull
-        assertThat(taskSlot.captured.payload).isEqualTo(JOURNALPOST_DIGITALSØKNAD)
-        assertThat(taskSlot.captured.metadata.getProperty("callId")).isEqualTo("kanalReferanseId")
-        assertThat(taskSlot.captured.taskStepType).isEqualTo(OppdaterOgFerdigstillJournalpostTask.TASK_STEP_TYPE)
-
-
-        verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
+        assertThat(slot.captured).isNotNull
+        assertThat(slot.captured.offset).isEqualTo(OFFSET)
+        assertThat(slot.captured.hendelseId).isEqualTo(HENDELSE_ID.toString())
+        assertThat(slot.captured.consumer).isEqualTo(HendelseConsumer.JOURNAL)
+        assertThat(slot.captured.metadata["journalpostId"]).isEqualTo(JOURNALPOST_ID)
+        assertThat(slot.captured.metadata["hendelsesType"]).isEqualTo("MidlertidigJournalført")
     }
+
 
     @Test
     fun `Ikke gyldige hendelsetyper skal ignoreres`() {
-        val record = opprettRecord(JOURNALPOST_PAPIRSØKNAD, "UgyldigType")
-
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
+        val ugyldigHendelsetypeRecord = opprettRecord(hendelseType = "UgyldigType")
+        val consumerRecord = ConsumerRecord("topic", 1, OFFSET, HENDELSE_ID, ugyldigHendelsetypeRecord)
 
         consumer.listen(consumerRecord, ack)
 
 
         verify(exactly = 0) {
-            mockTaskRepository.save(any())
+            mockJournalhendelseService.behandleJournalhendelse(any())
         }
-
         verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
+        val slot = slot<Hendelseslogg>()
+        verify(exactly = 1) {
+            mockHendelsesloggRepository.save(capture(slot))
+        }
+        assertThat(slot.captured).isNotNull
+        assertThat(slot.captured.offset).isEqualTo(OFFSET)
+        assertThat(slot.captured.hendelseId).isEqualTo(HENDELSE_ID.toString())
+        assertThat(slot.captured.consumer).isEqualTo(HendelseConsumer.JOURNAL)
+        assertThat(slot.captured.metadata["journalpostId"]).isEqualTo(JOURNALPOST_ID)
+        assertThat(slot.captured.metadata["hendelsesType"]).isEqualTo("UgyldigType")
     }
 
     @Test
     fun `Hendelser hvor journalpost ikke har tema for Barnetrygd skal ignoreres`() {
-        val record = opprettRecord(JOURNALPOST_IKKE_BARNETRYGD)
+        val ukjentTemaRecord = opprettRecord(temaNytt = "UKJ")
 
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
-
-        consumer.listen(consumerRecord, ack)
-
-
-        verify(exactly = 0) {
-            mockTaskRepository.save(any())
-        }
-
-        verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
-    }
-
-    @Test
-    fun `Hendelser hvor journalpost er alt FERDIGSTILT skal ignoreres`() {
-        val record = opprettRecord(JOURNALPOST_FERDIGSTILT)
-
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
+        val consumerRecord = ConsumerRecord("topic", 1, OFFSET, HENDELSE_ID, ukjentTemaRecord)
 
         consumer.listen(consumerRecord, ack)
 
 
         verify(exactly = 0) {
-            mockTaskRepository.save(any())
+            mockJournalhendelseService.behandleJournalhendelse(any())
         }
 
         verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
-    }
-
-    @Test
-    fun `Utgående journalposter skal ignoreres`() {
-        val record = opprettRecord(JOURNALPOST_UTGÅENDE_DOKUMENT)
-
-        val consumerRecord = ConsumerRecord("topic", 1, 1, 1, record)
-
-        consumer.listen(consumerRecord, ack)
-
-
-        verify(exactly = 0) {
-            mockTaskRepository.save(any())
+        val slot = slot<Hendelseslogg>()
+        verify(exactly = 1) {
+            mockHendelsesloggRepository.save(capture(slot))
         }
-
-        verify { ack.acknowledge() }
-        verify(exactly = 1) { mockHendelsesloggRepository.save(any()) }
+        assertThat(slot.captured).isNotNull
+        assertThat(slot.captured.offset).isEqualTo(OFFSET)
+        assertThat(slot.captured.hendelseId).isEqualTo(HENDELSE_ID.toString())
+        assertThat(slot.captured.consumer).isEqualTo(HendelseConsumer.JOURNAL)
+        assertThat(slot.captured.metadata["journalpostId"]).isEqualTo(JOURNALPOST_ID)
+        assertThat(slot.captured.metadata["hendelsesType"]).isEqualTo("MidlertidigJournalført")
     }
 
 
-    private fun opprettRecord(journalpostId: String,
+    private fun opprettRecord(temaNytt: String = "BAR",
                               hendelseType: String = "MidlertidigJournalført"): JournalfoeringHendelseRecord {
-        return JournalfoeringHendelseRecord("hendelseId",
+        return JournalfoeringHendelseRecord(HENDELSE_ID.toString(),
                                             1,
                                             hendelseType,
-                                            journalpostId.toLong(),
+                                            JOURNALPOST_ID.toLong(),
                                             "M",
                                             "BAR",
-                                            "BAR",
+                                            temaNytt,
                                             "SKAN_NETS",
                                             "kanalReferanseId",
                                             "BAR")
     }
 
     companion object {
-        const val JOURNALPOST_PAPIRSØKNAD = "111"
-        const val JOURNALPOST_DIGITALSØKNAD = "222"
-        const val JOURNALPOST_UTGÅENDE_DOKUMENT = "333"
-        const val JOURNALPOST_IKKE_BARNETRYGD = "444"
-        const val JOURNALPOST_FERDIGSTILT = "555"
+        const val OFFSET = 21L
+        const val HENDELSE_ID = 42L
+        const val JOURNALPOST_ID = "1"
+
+
     }
 }
